Bug 515913: [Model Explorer] Toggling advanced mode does not respect "rank" order
Implement a proper and complete delegating list in for customizations,
at the very least one that doesn't cause IndexOutOfBoundsExceptions by
reporting wrong results for size() and isEmpty().

When toggling advanced mode off by re-enabling the SimpleUML
customization, insert in the list in something like the place where
it belongs according to "rank" ordering.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=515913

Change-Id: Ife845cf766a0641b252e371492d2548d114b5672