Bug 509742: [Model Explorer][Usability] Diagrams owned by an element shall be located first for easier access
Define the implicit 'rank' of native model features as zero for
the purposes of ordering relative to customization features, which
have explicit ranks.  Use this in the content provider to insert
tree content from native model features in the appropriate relative
order and ensure that in new workspaces diagrams and tables are
presented before anything else.  Users can manually re-order
customizations in existing workspaces.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=509742

Change-Id: I16124412d74af4b163baeb5d4144d5c171eabd5e