Bug 528343: [I18N] Preference checking overhead for models that don't use i18n
Rework the InternationalizationUMLItemProviderAdapterFactory to
provide standard UML item-providers for resource sets that don’t
enabled i18n support, until such time as it is needed.  At that point,
dispose the standard item-providers and let the i18n-enabled ones
replace them.  An new event-listener protocol lets the item-provider
detect this change in i18n status.

Also improve performance of the i18n utility's tracking of
preference stores and fix the hash code calculation of the
PapyrusProjectScope (as it matters now that it is used as the
key in a map).

https://bugs.eclipse.org/bugs/show_bug.cgi?id=528343

Change-Id: Ifb415da65d6af26e5bfaac0ea3f1b9718b758227