Bug 532299 - Add tooltips on properties in architecture framework editor.
Added descriptions on all properties that show in the architecture
editor. Those descriptions appear in the status bar of the Eclipse
window when the property is selected (not as toolips when hovering over
properties). Also, some properties are categorized in the property sheet
into sub categories to ease understanding.

I also added the same descriptions as documentation of the AF model API.


Change-Id: I2a4e928b177ae5434c6f38255ea1bba41c822416
Signed-off-by: Maged Elaasar <melaasar@gmail.com>