Bug 533673: [Sequence Diagram] CombinedFragment should be created with one InteractionOperand
Add an edit-helper advice to resize the default operand of a new
combined fragment to fill the available space in the combined
fragment.  As this needs to use the edit-part controlling the
operand's presentation to obtain a command that resizes the operand
and captures the interaction fragments within its bounds, the
execution of this advice command is deferred and captured back into
the original context for undo/redo.

Change-Id: Ifebab84b5b2181bea36d1cd60bf1f9f61adc4fe3
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>