Bug 533676: [Sequence Diagram] Validation should be triggered after creation of CombinedFragment or InteractionOperand
Add custom well-formedness rules to check that messages and
execution specifications do not cross interaction operand boundaries.

Ensure that re-sizing of an interaction operand triggers validation of
that operand and the interaction fragments that it contains, plus any
messages originating or terminating within it.  Because the default
operand of a new combined fragment is now resized upon creation,
a new operand is likewise validated for its contents.

Fix validation hook and diagnostician extension point schema definitions
to match reality of the extension point implementations.

Remove UI contributions for validation, to avoid workflow interrution.
Validation framework is still slow, but less intrusive for the user. A
preference may control the validation activation on user actions.

Update the failing compilation on gmfdiag.common.tests.

Add a preference to switch on or off the validation.

Change-Id: Ic7d83eb18315ce714dc7c71bff38276f330ef260
Also-by: Rémi Schnekenburger <rschnekenburger@eclipsesource.com>
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>