Bug 533673: [Sequence Diagram] CombinedFragment should be created with one InteractionOperand
Add an edit-helper advice to resize the default operand of a new
combined fragment to fill the available space in the combined
fragment.  As this needs to use the edit-part controlling the
operand's presentation to obtain a command that resizes the operand
and captures the interaction fragments within its bounds, the
execution of this advice command is deferred and captured back into
the original context for undo/redo.

Change-Id: I774c025bf1534b54d1759454a24101aafccff2f2
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>
Signed-off-by: Nicolas FAUVERGUE <nicolas.fauvergue@cea.fr>