Bug 533679: Creation of CFrag should not graphically move elements
Fix the handling of operand ownership for execution specifications
to account for (a) that they are not started/finished only by
ExecutionOccurrenceSpecifications specifically and (b) that they
are not always provided by the rows of the grid model.

Since the updating of the rows and columns and fragment ownership
is so often done in a read-only context (following operation
execution on the history) we can avoid redundant calculations
by just always posting the update asynchronously and throttling
multiple pending calculations.

Altogether this results in the calculation of operand ownership
upon initial creation of a combined fragment not being immediately
overridden by an incorrect calculation that puts all of the
fragments back into the root interaction.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=533679

Change-Id: I2abb81a920e2a8a7f20d3a6b702d3194d8e7ae8e
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>