Bug 533683: [Sequence Diagram] Deletion of the last InteractionOperand
In the diagram, deletion of a multiple selection obtains a deletion
command separately from each selected edit-part, entirely within its
own context.  There is no reasonable way to share a single
InteractionContainerDeletionContext amongst these separate requests
that will not break other situations, so instead of pre-computing
the advice command that deletes the CombinedFragment when it will be
left without any operands, this decision is deferred to the actual
point of execution of an after advice.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=533683

Change-Id: I64b214df5540346c769c1ab769d914b279178e4c