Bug 533675: [Sequence Diagram] CombinedFragment can be created in any case in the diagram
Test creation of combined fragments
- covering nothing at all (in blank space)
- partially covering a message
- within another combined fragment

Also, update the tests for bug 535097 to use the editor fixture to
create interaction operands by automation of the creation tool, to
better simulate user interaction.  This requires explicitly disabling
snap to grid (which is on by default in Photon) so that the mouse
pointer can be placed as precisely as is required by the tests.

Change-Id: Ic68e6c49746d63dc3e7c93ad92bd89ab68937614
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>
Signed-off-by: Nicolas FAUVERGUE <nicolas.fauvergue@cea.fr>