Bug 507479: [SequenceDiagram] Choose lifelines covered by a combined fragment
Implement properties view for the InteractionFragment::covered
property of combined fragments and interaction operands. Restrict
the selection of covered lifelines to lifelines in the same interaction.

Add warning constraints for consistency of lifelines covered by
- interaction operand as compared to lifelines covered by the
  fragments that it owns
- combined fragment as compared to lifelines covered by fragments
  of its operands

Update the automatic validation (if the preference is set) of an
interaction operand that has has its owned fragments recomputed to
encompass the entire containing combined fragment, to handle the
impact on its lifeline coverage consistency constraint.

Update headers & plugin version

Change-Id: I340c816a3d62b38b2e811daa90b4d5c0c1fb5493
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>