Bug 536378: [RCPTT] Sequence diagram functional test regressions
- remove test cases from the "suites" folder that were redundant
  with identically-named test cases in the "tests" folder
- us CSS to select Roboto 11pt font in all tests for consistent
  (in some degree) geometry across platforms
- define ECL procedures to express assertions of the geometry of
  shapes in the diagram more concisely and with tolerances for
  platform variability (and perhaps other sources of variability)
- update test assertions to reflect current layout results on
  Mac platform with the intention that tolerances account for
  deviation in Linux build environment
- update stylesheet for labels

Note that some of the layout results as asserted now in the tests
may be functionally incorrect and so not flagged as such by tests
that should be failing. These will have to be reviewed once the
noise of test failures for reasons of invalid test construction
(e.g. asserting fractional anchors to 10 decimal places!) have been
successfully cleaned up and stabilized.

Change-Id: I0c757d3bd65ca30205a3cb1575dc61f81331a3b5
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>