Bug 537571: [Sequence Diagram] Time Constraints and Observations
Implement the TimeConstraint for DestructionOccurrenceSpecifications.

Implement the TimeObservation for ExecutionSpecifications and
DestructionOccurrenceSpecifications.

Implement time observations and constraints on message ends,
on lifelines and on execution specifications, as well as on the
lifeline head for creation message timing.

Ensure accurate creation feed-back for time constraints and
observations with an oversized shadow figure for clarity.

JUnit test coverage for creation of time observations
and time constraints in the diagram.

Change-Id: I57761c0b4819e7c0b2324bebb9f192a0aa495fef
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>