Bug 535696: [UsecaseDiagram] Strategy does not work when move use case in subject
Ensure that the drop strategy for use cases in subjects does not
duplicate the view when moving an existing use case within the subject
or when moving an existing use case from outside of the subject into
the subject.  This also covers the case of dropping a use case from
the model explorer into a subject that is already presenting that
use case within it, but in that case the drop strategy is just
skipped and does not participate further in the drop gesture.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>
Change-Id: Ic0f643b1da9aacbbba03a2f2673c93753aceb3c6
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>